criação de agendamentos, criação de status tanto para o adm quanto para o aluno e historico de agendamentos para o aluno finalizados.

+++ routes/web.php
@@ -31,16 +31,18 @@ Route::middleware('auth')->group(function () {
     Route::prefix('painel')->group(function () {
         Route::get('/', [PainelController::class, 'index'])->name('painel-index');
         Route::prefix('usuarios')->group(function () {
-            Route::get('/', [UsuarioController::class, 'index'])->name('usuarios-index');
-            Route::get('/create', [UsuarioController::class, 'create'])->name('usuarios-create');
-            Route::post('/', [UsuarioController::class, 'store'])->name('usuarios-store');
-            Route::get('/{id}/edit', [UsuarioController::class, 'edit'])->where('id', '[0-9]+')->name('usuarios-edit');
-            Route::put('/{id}', [UsuarioController::class, 'update'])->where('id', '[0-9]+')->name('usuarios-update');
-            Route::put('/senha/{id}', [UsuarioController::class, 'updateSenha'])->where('id', '[0-9]+')->name('usuarios-updateSenha');
-            Route::delete('/{id}', [UsuarioController::class, 'destroy'])->where('id', '[0-9]+')->name('usuarios-destroy');
+            Route::get('/', [UsuarioController::class, 'index'])->name('usuarios-index')->middleware('can:admin-access');
+            Route::get('/create', [UsuarioController::class, 'create'])->name('usuarios-create')->middleware('can:admin-access');
+            Route::post('/', [UsuarioController::class, 'store'])->name('usuarios-store')->middleware('can:admin-access');
+            Route::get('/{id}/edit', [UsuarioController::class, 'edit'])->where('id', '[0-9]+')->name('usuarios-edit')->middleware('can:admin-access');
+            Route::put('/{id}', [UsuarioController::class, 'update'])->where('id', '[0-9]+')->name('usuarios-update')->middleware('can:admin-access');
+            Route::put('/senha/{id}', [UsuarioController::class, 'updateSenha'])->where('id', '[0-9]+')->name('usuarios-updateSenha')->middleware('can:admin-access');
+            Route::delete('/{id}', [UsuarioController::class, 'destroy'])->where('id', '[0-9]+')->name('usuarios-destroy')->middleware('can:admin-access');
         });
         Route::prefix('agendamentos')->group(function () {
             Route::get('/', [AgendamentoController::class, 'index'])->name('agendamentos-index');
+            Route::get('/historicoAluno', [AgendamentoController::class, 'histAluno'])->name('agendamentos-histAluno');
+            Route::get('/historicoAluno', [AgendamentoController::class, 'index'])->name('agendamentos-parametros');
             Route::get('/create', [AgendamentoController::class, 'create'])->name('agendamentos-create');
             Route::post('/', [AgendamentoController::class, 'store'])->name('agendamentos-store');
             Route::get('/{id}/edit', [AgendamentoController::class, 'edit'])->where('id', '[0-9]+')->name('agendamentos-edit');
